Other version of playerController
This controller is much more effective then the previous one. The player movements are set using the rigidbody's velocity. This enables to have smooth transitions between ground and air movement with a minimum number of lines of code. 
The collision detection is still based on the first collider attached to the player object but this is easily to correct.